fix(config): Moves load-config into a separate file outside the lib scope
