+++ src/config/metadata.ts
@@ -1,10 +1,5 @@
-import * as cosmiconfig from "cosmiconfig";
 import { RegalError } from "../error";
 import { GameOptions } from "./game-options";
-
-const explorer = cosmiconfig("regal", {
-    searchPlaces: ["package.json", "regal.json"]
-});
 
 /**
  * Metadata about the game, such as its title and author.
@@ -32,55 +27,26 @@ export interface GameMetadata {
     options: Partial<GameOptions>;
 }
 
-const metadataOptionalStrings = [
-    "author",
-    "headline",
-    "description",
-    "homepage",
-    "repository"
-];
-
 export class MetadataManager {
     public static configLocation: string;
 
     public static getMetadata(): GameMetadata {
-        return MetadataManager._retrievalFunction();
+        if (MetadataManager._metadata === undefined) {
+            throw new RegalError(
+                "Metadata is not defined. Did you remember to load the config?"
+            );
+        }
+
+        return MetadataManager._metadata;
     }
 
-    public static forceConfig(config: GameMetadata): void {
-        MetadataManager._retrievalFunction = () => config;
+    public static setMetadata(metadata: GameMetadata): void {
+        MetadataManager._metadata = metadata;
     }
 
     public static reset(): void {
-        MetadataManager._retrievalFunction = MetadataManager._accessConfigFile;
-        MetadataManager.configLocation = undefined;
+        MetadataManager._metadata = undefined;
     }
 
-    /* tslint:disable:member-ordering */
-
-    private static _accessConfigFile = (): GameMetadata => {
-        const result =
-            MetadataManager.configLocation === undefined
-                ? explorer.searchSync()
-                : explorer.searchSync(MetadataManager.configLocation);
-
-        if (result.isEmpty) {
-            throw new RegalError("No metadata could be found for the game.");
-        }
-
-        const metadata: GameMetadata = {
-            name: result.config.name as string,
-            options: result.config.options as Partial<GameOptions>
-        };
-
-        metadataOptionalStrings.forEach(key => {
-            if (result.config[key] !== undefined) {
-                metadata[key] = result.config[key];
-            }
-        });
-
-        return metadata;
-    };
-
-    private static _retrievalFunction = MetadataManager._accessConfigFile;
+    private static _metadata: GameMetadata;
 }
